requirements.txt
# Core orchestration
langgraph>=0.3.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# LLM Providers
openai>=1.50.0

# Web search tools
tavily-python>=0.5.0

# Data models & validation
pydantic>=2.9.0

# Azure persistence
azure-cosmos>=4.8.0
azure-search-documents>=11.6.0
azure-identity>=1.18.0

# Document processing
pypdf>=5.0.0
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Vector embeddings
tiktoken>=0.8.0

# Environment & config
python-dotenv>=1.0.0

# Observability
langsmith>=0.2.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
